refactor: move pii-bouncer instructions out of the wizard
Addresses review feedback: the agent instructions (scan steps, abort
signals, report format) were hardcoded in the program's customPrompt —
"packaged English" that belongs in a context-mill skill, not the wizard.

- Collapse index.ts to a createSkillProgram() factory call; customPrompt
  is now a one-liner. The skill (loaded via skillId) drives the run.
- Delete wizard-side detection (detectPiiBouncerPrerequisites + the
  duplicated package walker a reviewer flagged). The skill detects
  prerequisites and emits [ABORT] signals; the wizard just routes them.
- Rename detect.ts -> abort-cases.ts, keeping only PII_BOUNCER_ABORT_CASES
  (terminal UX copy). The match regexes are the contract with the skill.

Wizard = engine + UX; context-mill skill = instructions. No product
knowledge left in the program.
